Fixed a bug in fits.h --
I was checking for (info.datatype < 0) to test that datatype was set
but some datatypes are defined in cfitsio.h as negative! Well, one
at least, FLOAT=-32. Rather than rely on the *value* of datatype to
determine if it has been set, I have now added a boolean "type_set"
which is false by default and set true when datatype has been defined.

dh

	modified:   include/common.h
	modified:   include/fits.h