Added get_system_date() to utilities, then in Common::imdir() automatically append this date to the imdir, creating if necessary.
Modify the UDP socket so that the user can opt out of using it by
setting ASYNCGROUP=none in the .cfg file. The async_main thread continues
to run so that the queue doesn't have to keep track of the async socket
status, but nothing will be sent to the multicast socket if it wasn't created.

dh

	modified:   include/common.h
	modified:   include/network.h
	modified:   include/utilities.h
	modified:   src/common.cpp
	modified:   src/network.cpp
	modified:   src/server.cpp
	modified:   src/utilities.cpp