Changed get_fitsname so that it returns a long (for error) instead of string, and the filename is returned by passing a reference to a string.
Also, move the insertion of the date subdirectory into get_fitsname()
so that it's called each time a new filename is generated, to make
sure that the correct date subdir is obtained.

This was done only for src/archon.cpp so far and will still need to
be done for astrocam.

dh

	modified:   include/common.h
	modified:   src/archon.cpp
	modified:   src/common.cpp