added optional linking of libcurl
This software does not use cURL but it does use cfitsio and cfitsio
can support cURL. If a system's cfitsio was built with cURL support
then it has to be linked here. If libcurl is found then link it,
but if not found then proceed without error.

This change requires increasing the cmake required version
from 3.5 to 3.12

dh