beginning of emulator so far, Archon-only, and limited functionality
There are no changes to the detector controller software.
This implements a separate sever which behaves like the Archon.
To use it, in the .cfg file,
set EMULATORPORT=some port,
set ARCHON_PORT=the same port
set ARCHON_IP=localhost (or IP of host that will
                         run the emulator)

Run the emulator in its own terminal (specifying the .cfg file)
and run the cameraserver as normal, but now it will talk to the
emulator instead of the Archon.

dh

 On branch master
	modified:   CMakeLists.txt
	new file:   include/emulator-archon.h
	new file:   include/emulator-server.h
	new file:   src/emulator-archon.cpp
	new file:   src/emulator-server.cpp