Change how mode is used (for Archon-only) --
Instead of requiring the user set the mode, now
automatically set the mode to DEFAULT. This requires
only that the ACF programmer create a .modes file once;
the user no longer must remember to always set the mode.
This makes it less onerous for a project that doesn't
need to use the mode functionality.

The .modes file must contain (at minimum):

 [MODE_DEFAULT]
 ARCH:NUM_DETECT=x
 ARCH:HORI_AMPS=x
 ARCH:VERT_AMPS=x

dh

 Changes to be committed:
	modified:   include/archon.h
	modified:   include/common.h
	modified:   include/emulator-archon.h
	modified:   src/archon.cpp