Implement "longexposure" command (Archon-only)
When set to true the controller server assumes that the exposure time
unit is in seconds. When set to false the exposure time unit is assumed
to be in milliseconds.  This command requires compatible Archon firmware
(ACF) with a parameter named "longexposure" which will be set =1 when
longexposure is true, and =0 when longexposure is false. The ACF must also
have appropriate exposure delay subroutines.  The default start-up condition
of the server is longexposure=false

Other changes --
 * add checks to exptime value (previously could exceed allowed values)
 * sleep for 90% of exposure delay (instead of 80%) before polling Archon.
   (probably OK but made need to revisit after further testing)

dh

	modified:   CMakeLists.txt
	modified:   cameraserver.cfg
	modified:   include/archon.h
	modified:   include/common.h
	modified:   include/fits.h
	modified:   src/archon.cpp
	modified:   src/server.cpp